requirements.txt
pandas>=1.5.0
openai>=1.0.0
python-dotenv>=1.0.0
docling==2.43.0
tiktoken>=0.5.0
requests>=2.28.0
numpy==1.26.4
transformers>=4.20.0
rich>=13.7.0
